Colocando as funções JavaScript em arquivos mais específicos
As funções de manutenção estavam todas no mesmo arquivo, deixando o projeto bem desorganizado
