+++ src/diario/etapas/index.html
@@ -187,11 +187,17 @@
 
 		</div>
 	</div>
+
 	<!-- IMPORTAÇÕES -->
-	<!--Conexao Servlets-->
-	<script src="../../js/diario/etapas/jquery-3.4.1.js"></script>
+	<!-- Arquivos base -->
+	<script src="../../js/jquery-3.4.1.js"></script>
 	<script src="../../js/materialize.js"></script>
-	<script src="../../js/diario/etapas/conectaServlet.js"></script>
+	<script src="../../js/diario/etapas/utils.js"></script>
+	<!-- Arquivos que executam as funções -->
+	<script src="../../js/diario/etapas/adiciona.js"></script>
+	<script src="../../js/diario/etapas/consulta.js"></script>
+	<script src="../../js/diario/etapas/deleta.js"></script>
+	<script src="../../js/diario/etapas/edita.js"></script>
 </body>
 
 </html>
