+++ src/js/diario/etapas/utils.js
@@ -1,0 +1,12 @@
+M.AutoInit();
+
+function recarrega() {
+	location.reload();
+}
+
+function limpaInputs() {
+	document.getElementById('AdicionaAnoEtapa').value = "";
+	document.getElementById('AdicionaValorEtapa').value = "";
+	document.getElementById('editaAno').value = "";
+	document.getElementById('editaValor').value = "";
+}
